# apps/delivery-api/docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: CLEVER Route Server Delivery API
  version: 0.1.0
  license:
    name: UNLICENSED
    identifier: LicenseRef-UNLICENSED
  summary: CLEVER route server delivery API HTTP contract
  description: >-
    Machine-readable contract for CLEVER route server communication with current
    admin clients, legacy Shopify compatibility routes, native driver mobile
    clients, and operational health checks. Human behavior notes live beside
    this file in docs/api/*.md.
  x-contract-revision: 2026-05-14
  x-source-repository: EVNSolution/clever-route-server
servers:
  - url: http://localhost:3000
    description: Local development server
  - url: https://{deliveryServerHost}
    description: Deployed delivery server
    variables:
      deliveryServerHost:
        default: delivery.example.com
security: []
...
paths:
...
  /shopify/auth/token-exchange:
    post:
      tags: [Shopify Embedded Admin]
      summary: Exchange a Shopify session token and store Admin API token metadata
      operationId: exchangeShopifySessionToken
      x-source-file: src/routes/shopify-auth.routes.ts
      security:
        - ShopifySessionToken: []
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ShopifyTokenExchangeRequest'
            examples:
              expectedShop:
                value:
                  shopDomain: example.myshopify.com
      responses:
        '200':
          description: Token material was exchanged and stored without exposing plaintext tokens.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ShopifyTokenExchangeEnvelope'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '502':
          description: Shopify token exchange or token persistence failed.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
              examples:
                exchangeFailed:
                  value:
                    data: null
                    error:
                      code: SHOPIFY_TOKEN_EXCHANGE_FAILED
                      message: Shopify token exchange failed
  /shopify/webhooks:
    post:
      tags: [Shopify Webhooks]
      summary: Receive and idempotently record a Shopify webhook
      operationId: receiveShopifyWebhook
      x-source-file: src/routes/shopify-webhook.routes.ts
      security:
        - ShopifyWebhookHmac: []
      parameters:
        - $ref: '#/components/parameters/ShopifyTopicHeader'
        - $ref: '#/components/parameters/ShopifyShopDomainHeader'
        - $ref: '#/components/parameters/ShopifyWebhookIdHeader'
        - $ref: '#/components/parameters/ShopifyApiVersionHeader'
        - $ref: '#/components/parameters/ShopifyEventIdHeader'
        - $ref: '#/components/parameters/ShopifyTriggeredAtHeader'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true
      responses:
        '200':
          description: Duplicate webhook receipt was already recorded.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ShopifyWebhookEnvelope'
        '202':
          description: New webhook receipt was accepted for storage.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ShopifyWebhookEnvelope'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
components:
  securitySchemes:
    ShopifySessionToken:
      type: http
      scheme: bearer
      bearerFormat: Shopify App Bridge session JWT
      description: Server-verified Shopify embedded app session token.
    DriverAccessToken:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Server-issued short-lived driver JWT with audience clever-delivery-driver.
    ShopifyWebhookHmac:
      type: apiKey
      in: header
      name: X-Shopify-Hmac-Sha256
      description: Base64 HMAC generated by Shopify over the raw request body.
  parameters:
...
    ShopifyTopicHeader:
      name: X-Shopify-Topic
      in: header
      required: true
      schema:
        type: string
      example: orders/create
    ShopifyShopDomainHeader:
      name: X-Shopify-Shop-Domain
      in: header
      required: true
      schema:
        pattern: '^[a-z0-9][a-z0-9-]*\.myshopify\.com$'
        type: string
      example: example.myshopify.com
    ShopifyWebhookIdHeader:
      name: X-Shopify-Webhook-Id
      in: header
      required: true
      schema:
        type: string
    ShopifyApiVersionHeader:
      name: X-Shopify-API-Version
      in: header
      required: false
      schema:
        type: string
      example: 2026-04
    ShopifyEventIdHeader:
      name: X-Shopify-Event-Id
      in: header
      required: false
      schema:
        type: string
    ShopifyTriggeredAtHeader:
      name: X-Shopify-Triggered-At
      in: header
      required: false
      schema:
        format: date-time
        type: string
  responses:
    BadRequest:
      description: Invalid request payload, query, path, or headers.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
    Unauthorized:
      description: Missing or invalid bearer token or HMAC.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
...
  schemas:
    Error:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
        message:
          type: string
      additionalProperties: true
    ErrorEnvelope:
      type: object
      required: [data, error]
      properties:
        data:
          type: 'null'
        error:
          $ref: '#/components/schemas/Error'
...
    ShopifyTokenExchangeRequest:
      type: object
      properties:
        shopDomain:
          type: string
          minLength: 1
          description: Optional expected-shop guard for session token verification.
      additionalProperties: false
    ShopifyTokenExchangeEnvelope:
      type: object
      required: [data, error]
      properties:
        data:
          type: object
          required: [shopDomain, tokenScopes, tokenStored]
          properties:
            shopDomain:
              type: string
            tokenScopes:
              type: array
              items:
                type: string
            tokenStored:
              const: true
        error:
          type: 'null'
    ShopifyWebhookEnvelope:
      type: object
      required: [data, error]
      properties:
        data:
          type: object
          required: [duplicate, webhookId]
          properties:
            duplicate:
              type: boolean
            webhookId:
              type: string
        error:
          type: 'null'
...